models/trained/ssd/weapon_detection_SSD/requirements.txt
tensorflow>=2.5.0
opencv-python>=4.5.3
numpy>=1.19.5
pyyaml>=5.4.1
pillow>=8.3.1
matplotlib>=3.4.3
scikit-learn>=1.0.0 